Rename DefaultConverterContext to GenericContext and move it to the Context sub-namespace
