+++ tests/Converter/GenericConverterTest.php
@@ -6,7 +6,7 @@ namespace Neusta\ConverterBundle\Tests\Converter;
 
 use Neusta\ConverterBundle\Converter\Converter;
 use Neusta\ConverterBundle\Converter\GenericConverter;
-use Neusta\ConverterBundle\Converter\DefaultConverterContext;
+use Neusta\ConverterBundle\Converter\Context\GenericContext;
 use Neusta\ConverterBundle\Tests\Fixtures\Factory\PersonFactory;
 use Neusta\ConverterBundle\Tests\Fixtures\Model\Person;
 use Neusta\ConverterBundle\Tests\Fixtures\Model\User;
@@ -15,7 +15,7 @@ use PHPUnit\Framework\TestCase;
 
 class GenericConverterTest extends TestCase
 {
-    /** @var Converter<User, Person, DefaultConverterContext> */
+    /** @var Converter<User, Person, GenericContext> */
     private Converter $converter;
 
     protected function setUp(): void
@@ -37,7 +37,7 @@ class GenericConverterTest extends TestCase
     {
         // Test Fixture
         $source = (new User())->setFirstname('Max')->setLastname('Mustermann');
-        $ctx = (new DefaultConverterContext())->setValue('separator', ', ');
+        $ctx = (new GenericContext())->setValue('separator', ', ');
         // Test Execution
         $target = $this->converter->convert($source, $ctx);
         // Test Assertion
